Fix float32 underflow silently zeroing saved PSDs; add PSD sanity-plot script
Found while building a script to sanity-check the backfilled PSDs: real
strain PSD values in the sensitive band (~1e-46 to 1e-48) are below
float32's smallest representable magnitude (~1.4e-45), so casting to
float32 flushes them to exact zero. Verified against bilby's own aLIGO
design curve, independent of any local test scaffolding -- confirmed
99.9% of PSD bins underflow to zero, including 1113/1121 bins in the
20-300Hz band alone. This affects both get_clean_backgrounds.py's
inline PSD saving and get_segment_psds.py's backfill output -- switched
both to float64 (whitened-sample arrays, which are O(1) magnitude, are
unaffected and left at float32).

The just-completed CIT backfill (segment_psds/*.pkl, 5343 contexts)
predates this fix and needs to be regenerated -- its checkpoint/resume
logic will otherwise treat those zeroed contexts as already done and
skip them.

Also adds plot_segment_psds.py, which randomly samples and plots PSDs
per run/IFO (pooling O3a/O3b and O4a/O4b like generate_real_noise_
samples.py) against a rough aLIGO design-curve reference -- this is
what caught the bug above, using a synthetic PSD built from bilby's
own published design curve, not any prior pipeline output.

+++ scripts/get_clean_backgrounds.py
@@ -193,7 +193,7 @@ def main() -> None:
                         # windows sliced from it), not one per sample.
                         if new:
                             psd_gps_starts.append(context_start)
-                            psds.append(np.asarray(psd.value, dtype=np.float32))
+                            psds.append(np.asarray(psd.value, dtype=np.float64))
                             if psd_freqs is None:
                                 psd_freqs = np.asarray(psd.frequencies.value, dtype=np.float64)
 
@@ -221,7 +221,10 @@ def main() -> None:
                 # sample count) -- match a sample to its PSD via gps_starts.
                 'psd_gps_starts': np.array(psd_gps_starts, dtype=np.float64),
                 'psd_freqs':      psd_freqs if psd_freqs is not None else np.array([]),
-                'psds':           np.array(psds, dtype=np.float32),
+                # float64, not float32: real strain PSD values in the sensitive
+                # band are ~1e-46 to 1e-48, below float32's smallest representable
+                # magnitude (~1.4e-45) -- float32 silently flushes them to zero.
+                'psds':           np.array(psds, dtype=np.float64),
             }
             print(f"\n  Done: {len(backgrounds[run][ifo]['samples'])} samples  |  "
                   f"failed fetches: {failed}")
